Move database triggers to separate folder
