+++ functions/src/index.ts
@@ -1,3 +1,3 @@
-export * as model from './model';
+export * as model from './model-triggers';
 export * as cron from './cron';
 export * from './api';
